requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
bs4>=0.0.1
llama-cpp-python>=0.2.11
tqdm>=4.65.0
requests>=2.31.0 